requirements.txt
lyricsgenius~=3.0.1
spotipy==2.20.0
requests==2.28.1
matplotlib==3.6.1
millify~=0.1.1
Pillow==9.2.0
psutil==5.9.3
tqdm~=4.64.0
python-dotenv==0.21.0
asyncspotify~=0.12.1
aiohttp==3.8.3
py-cord==2.2.2
pyrate-limiter==2.8.3
yt-dlp==2022.10.4